requirements.txt
fastapi
uvicorn
dynaconf
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib
python-multipart
bcrypt